final-project-snotty-slimesters/requirements.txt
pytest==7.1.3
pytest-pycodestyle
requests
flask
testfixtures
pymodm
dnspython
Pillow
pandas
scipy
biosppy
py-ecg-detectors
matplotlib
